refactored the usage of configuration to be isolated more from the ide fixture for easy tracking, added configurations variable in the intellij class
