+++ src/conftest.py
@@ -1,1 +1,4 @@
-pytest_plugins = ["src.fixtures.ide_fixture"]
+pytest_plugins = [
+    "src.fixtures.ide",
+    "src.fixtures.configuration"
+]
