+++ src/models/IDE/Intellij.py
@@ -5,6 +5,8 @@ import subprocess
 import time
 
 from src.models.application import Application
+from src.models.configuration.configuration import Configuration
+from src.models.configuration.configurations_object import ConfigurationsObject
 
 
 class Intellij(Application):
@@ -19,6 +21,7 @@ class Intellij(Application):
             x_coordinate=110,
             y_coordinate=470,
         )
+        self.configurations = ConfigurationsObject()
 
     def get_ide_version(self, ide_directory):
         pattern = re.compile(r"\d{3}\.\d{4}\.\d{2}")
@@ -34,6 +37,26 @@ class Intellij(Application):
         self.press_keys("enter")
         time.sleep(1)
         self.press_keys("enter")
+
+    def delete_all_configurations(self):
+        if not self.configurations:
+            return
+
+        self.click(self.config_run_region)
+        self.press_keys("ctrl", "a")
+        self.click(action="right_click")
+        self.press_keys("up")
+        self.press_keys("enter")
+        time.sleep(1)
+        self.configurations = []
+
+    def create_configuration(self):
+        self.click(self.config_run_region)
+        self.click(action="right_click")
+        self.press_keys("down")
+        self.press_keys("enter")
+        time.sleep(1)
+        ConfigurationsObject.configurations.append(Configuration())
 
     def image_locator(self, locator):
         return f"image:{self.IMG_DIR}/intellij/{locator}"
